Make object attacks don't be blocked
Somethies, object attacks would be blocked by things like object
creation, moving and destruction.

The creation of a CombatManager, who manages the attacks out of
HumanPlayer (and allows for more attack logic reuse) allows that
attacks don't block until the attacked is out of attacker range or
attacked is dead.