Add input acceptance
Before, we would pass the event for anyone who knew how to handle it,
and stop there, even if the element didn't handled the event. Now, with
input acceptance, the event would be stopped only if it was accepted.

This was done as a way to implement a better input event forwarding