Add an abstract mesh representation in the logic part of the game
We need to store a mesh reference in the object, to ease referencing
and rendering, but this can get in our way when we need to decouple
the different systems in the game.

This commit adds an "mesh interface" class to the logic part of the
game, to represent a mesh. Honestly, this can be almost anything you
want, like a java.lang.Object, or a Go (lol no generics) interface{}