Start adding the new asset file loader
Currently only add the code that loads the file and organizes the
dependency order.

It uses libyaml for loading the asset list, so it has been loaded too.
Might need to check how to make CMake discover libyaml location

The asset manager was mostly disabled, so the graphical game will not
run yet, but the tests and server will