Add the new terrain renderer
It now respects the terrain type, and render the terrain according to
the type.

It creates a huge texture, at the size of a section. For now, it
measures 16x16 per square, which gives a 4096x4096 texture for the
whole terrain section. Might seem big, but it isn't, and a videocard
can handle bigger textures.

The huge texture was done to avoid setting the texture for each
vertex, or square.