Make movement time-dependant
Now movement is time dependent. Instead of moving the same quantity
every frame, it moves only every 100 ms.

The animation feels somewhat slow, but this will be fixed.

Addresses #5. (And I believe that it fixes it, since I didn't seen the
same thing happen on attacks. But I think this needs to be checked).