Start work on a new logic engine
Instead of everyone snooping the object list, we'll use an event
emitter and event listener, so every object can get the added and
deleted objects and make their own lists, with the exact information
they need.

This commit adds a basic object event emitter and listener, some event
types and make the game compilable with that.

It also removes the locatable object -- there was no sense, because
the base object already had a virtual mesh -- and removed that
property hash table -- no use, and made the code longer and more
difficult