Start redoing the cities and teams
Now, our teams are very basic: just a name and a list of ally and
enemy teams

This commit adds the base class structure, a method to get the
diplomacy between cities, and add tests for this.

Also, the new city manager will get the newly-added objects, the
player who created the object and set the player's city.