Add an event to signal city changing for an object
The city changing event sees, for each player, when said player adds a
new object, and sets the object city to that player's city. It signals
an event for other subsystems to know: for example, the graphical
subsystem may want to tint the object with the player's color.

It's not tested yet, but the test suite compiled. However, there are
no tests that test this yet.