Make the ObjectManager use shared_ptrs instead of unique_ptr
This might ease managing multiple references of the same objects in
multiple areas.

When the game object was deallocated, sometimes it could be used after
it, causing dangling pointers, and therefore segfaults.

The shared_pointers have a refcount. The object would be destroyed
only after the last co-owner used it.