Notify objects when the attacked is hurt or dead
The combat manager now can notify and set the attacked status to hurt
if it received damage in the last tick, or to dead if its health
reached 0.

Both notifications will be useful, for example, to play a death
animation, or a warning sound to the user if its unit get attacked.