Modify vertex registering class for the renderer
The renderer now will accept a VertexData and a VertexInfo,
separately. This is because they are different: VertexData belongs to
the animator, and VertexInfo belongs to the mesh.

Essentially, we adapted the renderer to receive vertices from the new
mesh structure

And we made the renderer compile with the new code