Fix mesh positioning
The mesh would not get in the correct position despite its
corresponding object would be, as demonstrated by the debug output.

It was simply the world (or model) matrix for the object not being set
in the shader. This commit sets it.

(TODO: I might need to set a default uniform variable set for shaders)

But the material for the tent doesn't show up yet.