Fix crashes due to an invalid texture size count
Sometimes, the log would show that the maximum texture size is 0x0,
when it was not. This was happening probably due to some bug in
OpenGL.

We now only request the texture size once and cache the results
because, you know, people usually do not insert a new videocard with
the computer on, so this value will not change.