Fix some clang build issues
Some files had issues building with clang, due to default constructor
and invalid hashables.

The first was fixed by removing the default constructor values, since
we specify width and height when initializing the gui manager

The second was because the compiler was not recognizing the std::string
type. Including the header solved