Readd camera movement
Restart parsing events again.

The camera movement events are fully parsed and even replicated to the
camera.

The object build is replicated only event-wise, it does not do
anything.

We add a logic part of camera to make it available from the logic
engine so we can centralize event processing in the player manager.