Make the dummy object push actions in a more accurate way
It is not exactly gametick-accurate, we internally store a variable for
the input ticks, the amount of time `generateInput()` was run, that is
once per input tick. It is input-tick-accurate, however.

Since the game tick and input tick might not always run intertwined,
like 2 input ticks per 1 logic tick, sometimes the ticks that the
actions will run will be different. This is not an issue, it only
meant that something slowed down your computer, probably the debug
information that was printed.

Of course, when I do replays, they will be tick-accurate.