experiments: add an GUI image view control
This image view control will receive a buffer containing image data, or
an image file, and will show that data on screen

Currently only the image buffer loading code is implemented. The file
one is not.

We can draw an stretched image, a centered image and an zoomed, or
scaled, image.