gui: add relative positioning
Now we can, again, set the position relative to the size of the screen
instead of an absolute pixel value.

For example, instead of saying you want to position the control on
pixel position (100, 100) in an 1280x720 screen, you can say (0.078,
0.14).

This makes the interface more adaptable to different resolutions.