Call the callbacks on the GUI manager, not on the controls
This allows us to destroy the control in the callback, like when we
destroy the "New Game" button in its callback.

Before, doing it would cause crashes and issues, because we were
running the callback inside the button code. When the button was
destroyed, everything in memory about it was forgotten. If we tried to
access it, the code would crash.