Fix the framerate lock
Instead of using the time that takes for the loop to continue, use
only the time that took for things to render, as a basis for the
framerate lock.

Before, it did not work, because it considered the sleep as frame
time. For some reason, this resulted in the framerate lock, instead of
locking to framerate X, locking to 2X.

Now it works.