Move most OpenGL specific files to its own directory
Also add an include guard, to tell if any OpenGL headers were
included.

This will be done to add a null renderer, that will probably write
only diagnostic data, and will be useful for some testing.

Also, the OpenGL ES renderer, that will not be much different from
OpenGL, but it will be different enough to have its own renderer,
because the function names will probably differ