Add initial support to OpenGL ES in the opengl renderer
Since both renderers are now mutually exclusive, because of header files
conflicts, we hide the OpenGL ES part of the renderer behind of a
`#define`

We need to manually specify the precision of the vector and float
variables in the GLSL ES fragment shaders.

Also, use a simpler enum value for the texture type, GL_RGBA,
that means the same as GL_RGBA8 (8 bits per channel). The only
diference is that, in core OpenGL, we can be a little bit more
specific

OpenGL ES does not support certain texture formats, so we do not use them
(in the case of SRGB) or convert them to a supported format (in the
case of BGR24, we convert to RGB24)