input: use the object size as a bounding box source, instead of mesh
Use the object size to build a bounding box, not the mesh vertices.
Using the mesh vertices will consume a lot of CPU to query those
vertices, not so much with the object size, which is only a pair
of values. This also might allow you to select things that do not
have a lot of visible mesh area, like a thin object.

The only thing we lose is being able to know the object height,
but this can be added

Also revert the last commit, because the bounding box cache seemed
buggy.