Use flatbuffers to serialize now
Now, we use the flatbuffers library to serialize the player inputs to
the disk, if we record it.

Note that it stopped serializing on each write, now it serializes only
on exit (on class destruction, to be more correct). This might be bad,
especially if we want to get the input at the moment of a crash.

Good thing is now we serialize every player action we use, with all
information belonging to it