Change the way the record flatbuffers are saved
We save the input information in chunks of FlatBuffer encoded data, to
allow us to save the input as it comes, and make the input saving work
even if the game crashes.

THis might help us troubleshoot what happened to the user, and if it
has something to do with the logic subsystem.