Make the game save the object checksums in the input file
The object checksums are important; they will tell if the inputs will
run correctly, because the object properties need to be the same when
recording and reproducing the input. If they differ, same inputs can
yield different results.

We also check the object checksum when loading the input data, too.