logic: add a new pathfinder and a new object path manager
This new pathfinder does not leak memory (according to the address
sanitizer), and has code much more comprehensible and more clean than
its predecessor.

We use a lot of unique_ptrs to achieve that non-leaked memory, so data
locality might be bad. We can fix this by creating a pool of path
nodes.

Also, we add a new object path manager, with better support for
repathing moving objects, for example