logic: fix issues with the obstacle bitmap being set more than once
Sometimes, the obstacle bitmap will be set more than once, making hard
to see if more than one object set it. Now, we keep a count of how
many times it was set, and decrement it on each unset.

This solves the problem of two objects not recognizing each other if
they are too close