Add tests and dependencies for a new texture manager
This texture manager will be more modular than the current one,
allowing texture binding independent of the underlying graphical API.

Also, we will stop use devIL and use SDL_image, because it seems more
multithread-friendly and clean.

We add only testing, and the manager is in the same file as the tests
for now.

We also add a dummy texture environment (where you will adapt the
texture upload and bind for your API), to use with tests