graphics: make the device create the texture environment
More logical; now we let the device manage everything.

The only difference is that we return the texture environment as a
unique pointer rather than a raw pointer, because we will give
ownership to it to the texture manager, and let the manager manage the
environment lifetime.

For now, the device does not need to worry too much about texture
memory nor allocation rules. When this is needed, the pointer
type might change