Add attack events needed for the attack system refactor
The attack system will be reworked, to make easier to add
projectile-based attacks. Projectile attack start and damage will not
happen in the same tick, so some attack events will be added to make
the attack more or less async.

Also, because the attack might not even happen (miss), an attack miss
event is added for that.