build: refactor the cmake build file to improve readability
The CMakeLists.txt file was very disorganized, with a lot of repeated
code and statements, and a lot of things in one single file, so I split
them into different files.

Also, I removed the requirement for a googletest submodule, because we
now download the googletest data instead of cloning, to allow an easier
setup.